Refactor to use Config and return (#9)
Problem:
front end side we need to be able to see all of the results like per ticket cost, who admin is, and what the house fee is. this was it allows us to form some of the logic.

Solution:
update and return a config object. this object is immutable post contract creation.
---------

Co-authored-by: James <11054922+entrancedjames@users.noreply.github.com>